claim_checkworthiness/pyproject.toml
[project]
name = "claim-checkworthiness"
version = "1.0.0"
description = "Explainable Claim Checkworthiness Assessment for IJCAI 2026"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    # LLM APIs
    "openai>=1.0.0",
    "anthropic>=0.75.0",
    "together>=1.0.0",
    "groq>=0.9.0",

    # LLM Frameworks
    "dspy>=2.6.0",
    "baml-py>=0.70.0",

    # ML & Transformers
    "sentence-transformers>=5.2.0",
    "transformers>=4.57.3",
    "accelerate>=1.12.0",
    "datasets>=4.4.2",
    "scikit-learn>=1.8.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "catboost>=1.2.0",
    "imbalanced-learn>=0.14.1",

    # Data Processing
    "polars>=1.36.1",
    "pyarrow>=22.0.0",
    "numpy<2",

    # Evaluation & Metrics
    "nltk>=3.9.2",

    # Experiment Tracking
    "mlflow>=3.7.0",
    "optuna>=3.5.0",

    # Text Processing
    "contractions>=0.1.73",
    "emoji>=2.15.0",
    "ftfy>=6.3.1",
    "spacy>=3.8.11",
    "langdetect>=1.0.9",
    "wordsegment>=1.3.1",
    "sentencepiece>=0.2.1",

    # Utilities
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.2.1",
    "tqdm>=4.66.0",
    "json-repair>=0.54.3",
    "joblib>=1.5.2",
]

[project.optional-dependencies]
finetune = [
    "peft>=0.10.0",
    "bitsandbytes>=0.43.0",
]
visualization = [
    "plotly>=6.5.0",
    "umap-learn>=0.5.9.post2",
    "shap>=0.45.0",
]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"scripts/**/*.py" = ["E402"]
